streamlit/requirements.txt
streamlit>=1.39.0
plotly>=5.24.0
altair>=5.3.0
scikit-learn>=1.5.0
pandas>=1.5.0
pyarrow>=11.0.0
matplotlib>=3.5.0
numpy>=1.21.0
joblib>=1.1.0
pyyaml>=6.0
